refactor(slack): extract gate & mention policy into SlackGateMixin (adapter.py god-file slice R5)
